requirements.txt
psycopg2-binary==2.9.9
fuzzywuzzy==0.18.0
mygene==3.2.2
PyPDF2==3.0.1
requests==2.32.3
unittest-mock